requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
pillow